legacy/flight-fare-monitor/requirements.txt
requests
python-dotenv
pyyaml
sqlalchemy
psycopg2-binary
pandas
streamlit>=1.32.0
